feat(be): 재고 반영 이벤트에 대한 주문 도메인의 구독 처리

+++ backend/src/main/java/com/deliveranything/domain/order/event/sse/seller/OrderPaidForSellerEvent.java
@@ -20,7 +20,7 @@ public record OrderPaidForSellerEvent(
         order.getStore().getSellerProfileId(),
         order.getAddress(),
         order.getStoreNote(),
-        order.getTotalPrice().longValue()
+        order.getTotalPrice()
     );
   }
 }
